Convert Chicago to a split city.
This requires supporting split cities in general. A split city is one in
which the station slots are not all clustered together. This means having
a station in that city doesn't necessarily give a railroad access to all
paths through that space. This is represented in the JSON configs by
noting the capcity of each branch of the split.

In addition to creating this new type of object for board spaces and
placed tiles, it requires allowing the railroad CSV to indicate stations
on split city branches. This is achieved by supporting an arbitrary
number of fields at the end of each row, which are expected to be mappings
between a cell coordinate and a branch on that space.

It also means allowing the handling of edges to be more expressive, since
we can no longer assume all cities do paths the same. The sped is:
- edges should be a list
- each element of that list should be an int or list
- if it's a list, it represents a path. length 1 terminates, length 2 is
standard track, and length 3+ is a standard city.
- if it's an int, it's equivalent to a length 1 list; a terminating path.