Don't propagate stations to placed tiles.
When placing a new tile, the old tile neighbors would need to be mapped
to the new tile. This can be a bit tricky due to changing orientations,
and what it would mean for the neighbors. Rather than some gnarly logic,
we side-step the whole issue by insisting stations should only be placed
after all tiles.