Route phase management through the game object.
Since it's already present in many places, and will be needed in many
places, it's a natural place to store this info. Also a natural place to
load it, and provide other game specific information. And provides an
easy way to handle each games' different phase progression, as well as
comparing non-integer phases.