Setup structure for different route finding rules.
Currently only implements one: some games allow you to visit a town
without it counting against your route length. This is indicated in the
game JSON object, although it defaults to False.

A rule that impacts route finding it harder to implement as a hook, since
it requires a few modifications. Maybe some future rules will use hooks,
but this is another option.