Generalize private companies through the game object.
The biggest piece is giving each game a place it can define how its
private company logic, which is accessed through the game object. It also
adds a hook to allow modifying route values after they're calculated,
which is utilized by the mail contract.

Also removes all explicit references to the meat packing company and
steamboat company, moving instead to allowing private company tokens to
be placed. To define their values, a new properties field is introduced
to tiles and spaces which is intended for use by game-specific code.