Add and publicize private company info.
Namely, the valid coordinates for private companies which place tokens, and
where (if anywhere) those tokens should be placed before the company is owned
by a railroad.

This may get moved into a JSON config file at some point, but I need to se some
more private companies to determine the best way to do it.