Ensure home and reserved spaces are respected.
Tiles with a home station should ALWAYS make sure there's room for the
home station (or it's present). Tiles with a reserved station spot
should make sure there's room for the reserved station (or it's present)
as long as the railroad has not been removed from the game AND the phase
has not cancelled the reservation.

Also, if a railroad is in play and its home tile has too many stations
on it, raise the capacity error. If it's not in play, raise the home
slot error.