Split tiles.json into DB and game inventory.
Since all the games draw from the same pool of tiles, it makes sense to
centralize all the info about the tiles themselves. The DB file contains
all that info, and is a JSON file (despite its name). Each game will
continue to define a tiles.json file, but now it contains a mapping from
the ID of each tile in the game to the quantity of tles in the game. To
indicate a tile which has no limit on quantity (e.g. #7 in 1846), set it
to null.