Check train limit.
Confirm that each railroad adheres to the train limit set by the game
for the current phase. This includes defining a structure games can use
to describe their train limit by stages, including breaking it down by
train type. It also should be extensible for when various types of
railroads are required, as well as different limits for certain train
types.