Support split towns.
Create a SplitTown placetile and boardtile. In both cases, it identifies
the split town by its path groupings. If there's more than 1 grouping,
it's split.

Also add a new field to SplitCity and SplitTown called branches, which
just lists the branch paths. This is needed since SplitTown doesn't have
a concept of capcaity in which it can be encoded.