Generalize split city path handling.
A big part of this is detecting station inclusion when that station is
on a split city. Due to route generation considering a single 2-length
segment at a time, this needs to be done after the fact. Confirms all
the split city tile's neighbors in the path are all the branch's
neighbors (or nothing).

The other part is checking if the path from the entry cell to a
neighboring cell is passable. On a normal city, this is entirely
determined by if there are any open slots or the railroad's station is
present. For a split city, this is only relevant for the branch formed
by the entry and exit cells.