Unify private companies interface.
Hoist the common eleemnts out of the existing private_companies.py module
(for 1846) to make it easier to define private companies for new games.

Namely, handling token-base privates and the load and load_csv functions
are moved into routes18xx.games, and token classes are given a static
NAME field.