Extend routes to count towns when applicable.
In games which give trains the value for passing through towns but don't
count as having visited them, routes shouldn't end when visit hits zero.
There may be towns reachable before encountering another city/terminus
whose value can be counted.

In the algorithm design, there's no way to back out the route when we go
one city/terminus too far, nor lookahead to see such a space. Instead,
since the desired routes are added in the subroute step, we query the
train during the filter step for if the route's stops are valid.